+++ sclpl/project/__init__.py
@@ -2,5 +2,6 @@
 
 from sclpl.project.context import ProjectContext, discover, load
 from sclpl.project.identity import WorkflowIdentity, identify
+from sclpl.project.scripts import Registration
 
-__all__ = ["ProjectContext", "WorkflowIdentity", "discover", "identify", "load"]
+__all__ = ["ProjectContext", "Registration", "WorkflowIdentity", "discover", "identify", "load"]